Rename package to com.noseplugapp.android
