+++ android/app/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.2"
     defaultConfig {
-        applicationId "edu.gatech.cs.environmentalodors"
+        applicationId "com.noseplugapp.android"
         minSdkVersion 15
         targetSdkVersion 25
         versionCode 1
