+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="edu.gatech.cs.environmentalodors">
+    package="com.noseplugapp.android">
 
     <!--
          The ACCESS_COARSE/FINE_LOCATION permissions are not required to use
@@ -50,7 +50,7 @@
             android:name=".OdorEventDetailsActivity"
             android:label="@string/title_activity_odor_event_details"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".OdorReportDetailsActivity"></activity>
+        <activity android:name=".OdorReportDetailsActivity" />
     </application>
 
 </manifest>
