+++ android/app/src/main/java/com/noseplugapp/android/GoogleApiClientWrapper.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs.environmentalodors;
+package com.noseplugapp.android;
 
 import android.Manifest;
 import android.content.Context;
@@ -10,14 +10,13 @@ import android.support.annotation.Nullable;
 import android.support.v4.app.ActivityCompat;
 import android.util.Log;
 
+import com.noseplugapp.android.events.LocationEvent;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.maps.model.LatLng;
 
 import org.greenrobot.eventbus.EventBus;
-
-import edu.gatech.cs.environmentalodors.events.LocationEvent;
 
 /**
  * GoogleApiClientWrapper wraps the GoogleApiClient. This should help separate the activities from
