+++ android/app/src/main/java/com/noseplugapp/android/OdorReportDetailsActivity.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs.environmentalodors;
+package com.noseplugapp.android;
 
 import android.os.Bundle;
 import android.os.ParcelUuid;
@@ -6,11 +6,11 @@ import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.TextView;
 
+import com.noseplugapp.android.models.OdorReport;
+
 import java.util.UUID;
 
-import edu.gatech.cs.environmentalodors.models.OdorReport;
-
-import static edu.gatech.cs.environmentalodors.IntentExtraNames.ODOR_REPORT_ID;
+import static com.noseplugapp.android.IntentExtraNames.ODOR_REPORT_ID;
 
 public class OdorReportDetailsActivity extends AppCompatActivity {
