+++ android/app/src/main/java/com/noseplugapp/android/models/OdorReport.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs.environmentalodors.models;
+package com.noseplugapp.android.models;
 
 import com.google.android.gms.maps.model.LatLng;
