+++ android/app/src/main/res/layout/activity_maps.xml
@@ -11,7 +11,7 @@
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="edu.gatech.cs.environmentalodors.MapsActivity"
+        tools:context="com.noseplugapp.android.MapsActivity"
         />
 
     <android.support.design.widget.FloatingActionButton
