+++ android/app/src/main/res/layout/activity_odor_event_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="edu.gatech.cs.environmentalodors.OdorEventDetailsActivity">
+    tools:context="com.noseplugapp.android.OdorEventDetailsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
