+++ android/app/src/main/res/layout/activity_odor_report_details.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="edu.gatech.cs.environmentalodors.OdorReportDetailsActivity"
+    tools:context="com.noseplugapp.android.OdorReportDetailsActivity"
     >
 
     <TextView
