+++ android/app/src/main/res/layout/activity_report_form.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="edu.gatech.cs.environmentalodors.ReportFormDateTimeActivity">
+    tools:context="com.noseplugapp.android.ReportFormDateTimeActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
