+++ android/app/src/main/res/layout/activity_report_form_description.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="edu.gatech.cs.environmentalodors.ReportFormDescriptionActivity">
+    tools:context="com.noseplugapp.android.ReportFormDescriptionActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
