+++ android/app/src/main/res/layout/content_odor_event_details.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="edu.gatech.cs.environmentalodors.OdorEventDetailsActivity"
+    tools:context="com.noseplugapp.android.OdorEventDetailsActivity"
     tools:showIn="@layout/activity_odor_event_details">
 
     <ListView
