+++ android/app/src/main/res/layout/content_report_form_description.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="edu.gatech.cs.environmentalodors.ReportFormDescriptionActivity"
+    tools:context="com.noseplugapp.android.ReportFormDescriptionActivity"
     tools:showIn="@layout/activity_report_form_description">
 
     <Spinner
